Migrate logic code out from runner to be able to integrate junit 5
