+++ localstack/ext/java/src/main/java/cloud/localstack/LocalstackExtension.java
@@ -1,0 +1,17 @@
+package cloud.localstack;
+
+import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+/**
+ * Simple JUnit extension for JUnit 5.
+ *
+ * @author Patrick Allain
+ */
+public class LocalstackExtension implements BeforeTestExecutionCallback {
+
+    @Override
+    public void beforeTestExecution(final ExtensionContext context) {
+        Localstack.INSTANCE.setupInfrastructure();
+    }
+}
