+++ localstack/ext/java/src/main/java/cloud/localstack/docker/LocalstackDockerTestRunner.java
@@ -1,26 +1,11 @@
 package cloud.localstack.docker;
 
-import java.lang.annotation.Annotation;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.logging.Logger;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import lombok.Getter;
-import org.apache.commons.lang3.StringUtils;
+import cloud.localstack.Localstack;
+import cloud.localstack.docker.annotation.LocalstackDockerAnnotationProcessor;
+import cloud.localstack.docker.annotation.LocalstackDockerConfiguration;
 import org.junit.runner.notification.RunNotifier;
 import org.junit.runners.BlockJUnit4ClassRunner;
 import org.junit.runners.model.InitializationError;
-
-import cloud.localstack.LocalstackTestRunner;
-import cloud.localstack.ServiceName;
-import cloud.localstack.docker.annotation.IEnvironmentVariableProvider;
-import cloud.localstack.docker.annotation.IHostNameResolver;
-import cloud.localstack.docker.annotation.LocalstackDockerProperties;
-import cloud.localstack.docker.command.RegexStream;
 
 /**
  * JUnit test runner that automatically pulls and runs the latest localstack docker image
@@ -36,36 +21,26 @@ import cloud.localstack.docker.command.RegexStream;
  * by using the LocalstackDockerProperties annotation with an IHostNameResolver.
  *
  * @author Alan Bevier
+ * @author Patrick Allain
  */
 public class LocalstackDockerTestRunner extends BlockJUnit4ClassRunner {
 
-    private static final Logger LOG = Logger.getLogger(LocalstackDockerTestRunner.class.getName());
 
-    private static String externalHostName = "localhost";
-    private static boolean pullNewImage = true;
-    private static boolean randomizePorts = false;
-    private static Map<String, String> environmentVariables = new HashMap<>();
+    private final LocalstackDockerConfiguration dockerConfiguration;
 
-    @Getter
-    private static LocalstackDocker localstackDocker = LocalstackDocker.getLocalstackDocker();
+    private LocalstackDocker localstackDocker = LocalstackDocker.INSTANCE;
 
 
     public LocalstackDockerTestRunner(Class<?> klass) throws InitializationError {
         super(klass);
-        processAnnotations(klass.getAnnotations());
+        this.dockerConfiguration = new LocalstackDockerAnnotationProcessor(klass).process();
     }
 
     @Override
     public void run(RunNotifier notifier) {
-        LocalstackTestRunner.teardownInfrastructure();
-
-        localstackDocker.setExternalHostName(externalHostName);
-        localstackDocker.setPullNewImage(pullNewImage);
-        localstackDocker.setRandomizePorts(randomizePorts);
-        localstackDocker.setEnvironmentVariables(environmentVariables);
-
+        Localstack.teardownInfrastructure();
         try {
-            localstackDocker.startup();
+            localstackDocker.startup(dockerConfiguration);
             super.run(notifier);
         }
         finally {
@@ -73,42 +48,5 @@ public class LocalstackDockerTestRunner extends BlockJUnit4ClassRunner {
         }
     }
 
-    private void processAnnotations(Annotation[] annotations) {
-        for(Annotation annotation : annotations) {
-            if(annotation instanceof LocalstackDockerProperties) {
-                processDockerPropertiesAnnotation((LocalstackDockerProperties)annotation);
-            }
-        }
-    }
 
-
-    private void processDockerPropertiesAnnotation(LocalstackDockerProperties properties) {
-        pullNewImage = properties.pullNewImage();
-        randomizePorts = properties.randomizePorts();
-
-        try {
-            IHostNameResolver hostNameResolver = properties.hostNameResolver().newInstance();
-            String resolvedName = hostNameResolver.getHostName();
-            if(StringUtils.isNotBlank(resolvedName)) {
-                externalHostName = resolvedName;
-            }
-            LOG.info("External host name is set to: " + externalHostName);
-        }
-        catch(InstantiationException | IllegalAccessException ex) {
-            throw new IllegalStateException("Unable to resolve hostname", ex);
-        }
-
-        try {
-            IEnvironmentVariableProvider environmentProvider = properties.environmentVariableProvider().newInstance();
-            environmentVariables = environmentProvider.getEnvironmentVariables();
-        }
-        catch(InstantiationException | IllegalAccessException ex) {
-            throw new IllegalStateException("Unable to get environment variables", ex);
-        }
-
-        String services = String.join(",", properties.services());
-        if(StringUtils.isNotEmpty(services)) {
-            environmentVariables.put("SERVICES", services);
-        }
-    }
 }
