+++ localstack/ext/java/src/test/java/cloud/localstack/docker/DockerOnlySQSFunctionalityTest.java
@@ -106,7 +106,7 @@ public class DockerOnlySQSFunctionalityTest {
 
     private SQSConnection createSQSConnection() throws Exception {
         SQSConnectionFactory connectionFactory = SQSConnectionFactory.builder().withEndpoint(
-            LocalstackDockerTestRunner.getLocalstackDocker().getEndpointSQS()).withAWSCredentialsProvider(
+            LocalstackDocker.INSTANCE.getEndpointSQS()).withAWSCredentialsProvider(
                 new AWSStaticCredentialsProvider(TestUtils.TEST_CREDENTIALS)).build();
         return  connectionFactory.createConnection();
     }
