+++ localstack/ext/java/src/test/java/cloud/localstack/docker/LocalstackDockerTest.java
@@ -1,6 +1,7 @@
 package cloud.localstack.docker;
 
 import cloud.localstack.DockerTestUtils;
+import cloud.localstack.docker.annotation.LocalstackDockerConfiguration;
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.model.SendMessageResult;
@@ -8,6 +9,8 @@ import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+
+import java.util.Collections;
 
 import static org.junit.Assert.assertNotNull;
 
@@ -18,9 +21,14 @@ public class LocalstackDockerTest {
 
     @Test
     public void startup() {
-        LocalstackDocker localstackDocker = LocalstackDocker.getLocalstackDocker();
-        localstackDocker.setRandomizePorts(true);
-        localstackDocker.startup();
+        LocalstackDocker localstackDocker = LocalstackDocker.INSTANCE;
+        LocalstackDocker.INSTANCE.startup(new LocalstackDockerConfiguration(
+                false,
+                true,
+                "localhost",
+                Collections.emptyMap()
+
+        ));
 
         AmazonSQS amazonSQS = DockerTestUtils.getClientSQS();
         String queueUrl = amazonSQS.createQueue("test-queue").getQueueUrl();
@@ -33,16 +41,26 @@ public class LocalstackDockerTest {
 
         thrown.expect(IllegalStateException.class);
 
-        localstackDocker.startup();
+        LocalstackDocker.INSTANCE.startup(new LocalstackDockerConfiguration(
+                false,
+                true,
+                "localhost",
+                Collections.emptyMap()
+
+        ));
         localstackDocker.stop();
     }
 
     @Test
     public void stop() {
-        LocalstackDocker localstackDocker = LocalstackDocker.getLocalstackDocker();
-        localstackDocker.setRandomizePorts(true);
-        localstackDocker.startup();
-        localstackDocker.stop();
+        LocalstackDocker.INSTANCE.startup(new LocalstackDockerConfiguration(
+                false,
+                true,
+                "localhost",
+                Collections.emptyMap()
+
+        ));
+        LocalstackDocker.INSTANCE.stop();
 
         AmazonSQS amazonSQS = DockerTestUtils.getClientSQS();
         thrown.expect(SdkClientException.class);
@@ -51,6 +69,6 @@ public class LocalstackDockerTest {
 
     @After
     public void tearDown() {
-        LocalstackDocker.getLocalstackDocker().stop();
+        LocalstackDocker.INSTANCE.stop();
     }
 }
